Split RoomListPresenter and introduce HomePresenter
