+++ features/home/impl/src/main/kotlin/io/element/android/features/home/impl/HomeEvents.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 New Vector Ltd.
+ * Copyright 2025 New Vector Ltd.
  *
  * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
  * Please see LICENSE files in the repository root for full details.
@@ -7,4 +7,4 @@
 
 package io.element.android.features.home.impl
 
-internal fun RoomListState.contentAsRooms() = contentState as RoomListContentState.Rooms
+sealed interface HomeEvents
