+++ features/home/impl/src/main/kotlin/io/element/android/features/home/impl/HomeState.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2023, 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.features.home.impl
+
+import androidx.compose.runtime.Immutable
+import io.element.android.features.home.impl.roomlist.RoomListState
+import io.element.android.features.logout.api.direct.DirectLogoutState
+import io.element.android.libraries.designsystem.utils.snackbar.SnackbarMessage
+import io.element.android.libraries.matrix.api.user.MatrixUser
+
+@Immutable
+data class HomeState(
+    val matrixUser: MatrixUser,
+    val showAvatarIndicator: Boolean,
+    val hasNetworkConnection: Boolean,
+    val roomListState: RoomListState,
+    val snackbarMessage: SnackbarMessage?,
+    val canReportBug: Boolean,
+    val directLogoutState: DirectLogoutState,
+    val eventSink: (HomeEvents) -> Unit,
+) {
+    val displayActions = true
+}
