+++ features/home/impl/src/main/kotlin/io/element/android/features/home/impl/HomeStateProvider.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2025 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.features.home.impl
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import io.element.android.features.home.impl.roomlist.RoomListState
+import io.element.android.features.home.impl.roomlist.RoomListStateProvider
+import io.element.android.features.home.impl.roomlist.aRoomListState
+import io.element.android.features.logout.api.direct.DirectLogoutState
+import io.element.android.features.logout.api.direct.aDirectLogoutState
+import io.element.android.libraries.designsystem.utils.snackbar.SnackbarMessage
+import io.element.android.libraries.matrix.api.core.UserId
+import io.element.android.libraries.matrix.api.user.MatrixUser
+import io.element.android.libraries.ui.strings.CommonStrings
+
+open class HomeStateProvider : PreviewParameterProvider<HomeState> {
+    override val values: Sequence<HomeState>
+        get() = sequenceOf(
+            aHomeState(),
+            aHomeState(hasNetworkConnection = false),
+            aHomeState(snackbarMessage = SnackbarMessage(CommonStrings.common_verification_complete)),
+        ) + RoomListStateProvider().values.map {
+            aHomeState(roomListState = it)
+        }
+}
+
+internal fun aHomeState(
+    matrixUser: MatrixUser = MatrixUser(userId = UserId("@id:domain"), displayName = "User#1"),
+    showAvatarIndicator: Boolean = false,
+    hasNetworkConnection: Boolean = true,
+    snackbarMessage: SnackbarMessage? = null,
+    roomListState: RoomListState = aRoomListState(),
+    canReportBug: Boolean = true,
+    directLogoutState: DirectLogoutState = aDirectLogoutState(),
+    eventSink: (HomeEvents) -> Unit = {}
+) = HomeState(
+    matrixUser = matrixUser,
+    showAvatarIndicator = showAvatarIndicator,
+    hasNetworkConnection = hasNetworkConnection,
+    snackbarMessage = snackbarMessage,
+    canReportBug = canReportBug,
+    directLogoutState = directLogoutState,
+    roomListState = roomListState,
+    eventSink = eventSink,
+)
