+++ features/home/impl/src/main/kotlin/io/element/android/features/home/impl/di/RoomListModule.kt
@@ -12,6 +12,8 @@ import dagger.Binds
 import dagger.Module
 import io.element.android.features.home.impl.filters.RoomListFiltersPresenter
 import io.element.android.features.home.impl.filters.RoomListFiltersState
+import io.element.android.features.home.impl.roomlist.RoomListPresenter
+import io.element.android.features.home.impl.roomlist.RoomListState
 import io.element.android.features.home.impl.search.RoomListSearchPresenter
 import io.element.android.features.home.impl.search.RoomListSearchState
 import io.element.android.libraries.architecture.Presenter
@@ -20,6 +22,9 @@ import io.element.android.libraries.di.SessionScope
 @ContributesTo(SessionScope::class)
 @Module
 interface RoomListModule {
+    @Binds
+    fun bindRoomListPresenter(presenter: RoomListPresenter): Presenter<RoomListState>
+
     @Binds
     fun bindSearchPresenter(presenter: RoomListSearchPresenter): Presenter<RoomListSearchState>
